Updates (#2)
* improve text and explanations
* add absolute credit to icrc_notify() response
* add absolute credit to icrcX_deposit response
* turn all error variants into records for extendability
* replace X by 84
* rename files from icrcX to icrc-84
* Change arguments of icrc84_deposit

We now specify a `from` account which is a full ICRC-1 `Account`
type, not only a subaccount. The idea is that the caller can now
draw funds from accounts that he does not own himself. Assuming
the allowance is set up correctly, which means the spender account
must be unique to him.

* allow withdrawals directly to third parties